remove INodeMap

+++ hadoop-hdfs-project/hadoop-hdfs/src/main/java/org/apache/hadoop/hdfs/server/namenode/FSDirectory.java
@@ -1517,7 +1517,7 @@ public class FSDirectory implements Closeable {
   }
   
   @VisibleForTesting
-  int getInodeMapSize() {
+  long getInodeMapSize() {
     return inodeMap.size();
   }
